ai-103-text-speech/requirements.txt
azure-ai-projects>=2.0.1
azure-ai-textanalytics>=5.3.0
azure-ai-translation-text>=1.0.0
azure-cognitiveservices-speech>=1.46.0
azure-identity>=1.19.0
openai>=1.54.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
python-multipart>=0.0.17
